Move Log implementations to to utils.go
